+++ src/constants/cssAtRulesWithoutDeclarations/index.js
@@ -1,0 +1,5 @@
+import { cssAtRules } from '../cssAtRules';
+import { cssAtRulesWithDeclarations } from '../cssAtRulesWithDeclarations';
+import { difference } from '../../converters/difference';
+
+export const cssAtRulesWithoutDeclarations = difference(cssAtRules, cssAtRulesWithDeclarations);